requirements.txt
pandas
python-dateutil
matplotlib